+++ BicycleRental/src/main/java/com/wxh/bicyclerental/mapper/CouponMapper.java
@@ -7,10 +7,10 @@ import java.util.List;
 
 @Mapper
 public interface CouponMapper {
-    public Coupon selectOne(Integer couponId);
+    public Coupon selectOne(Long couponId);
     public List<Coupon> select();
     public int insert(Coupon coupon);
-    public int delete(Integer id);
+    public int delete(Long id);
     public int update(Coupon coupon);
     public List<Coupon> selectByState(Integer state);
 }
